split todo-item component
